refactor(integration): restructure ec app into layer-first layout
- Reorganize src from feature-first folders into entry/ (controllers,
  middleware, event handlers), usecase/, domain/, infra/db/, config/
- Move request schemas into domain/ so layer imports stay one-way
  (entry -> usecase -> infra; usecase reads value shapes from domain)
- Extract the user type into domain/ to keep config from importing entry
- Run vitest with the forks pool: native addons cannot register twice
  inside worker_threads

Verified: precommit-verify △ (849da55fc3b7e622)
