+++ integration/ec-backend/src/domain/auth.schema.ts
