+++ integration/ec-backend/src/domain/product.schema.ts
