+++ integration/ec-backend/src/domain/user.types.ts
@@ -1,0 +1,1 @@
+export type EcUser = { readonly id: number; readonly email: string };
