+++ integration/ec-backend/src/entry/middleware/logging.middleware.ts
