+++ integration/ec-backend/src/infra/db/schema.ts
